+++ rt_thread_3.1.3-ucosiii_3.03-wrapper/uCOS-III_Wrapper/uCOS-III/os_sem.c
@@ -504,10 +504,7 @@ OS_OBJ_QTY  OSSemPendAbort (OS_SEM  *p_sem,
        *p_err =  OS_ERR_PEND_ABORT_NONE;
         return ((OS_OBJ_QTY)0u);        
     }
-    else
-    {
-        CPU_CRITICAL_EXIT();
-    }
+    CPU_CRITICAL_EXIT();
     
     if(opt&OS_OPT_PEND_ABORT_ALL)
     {
@@ -518,7 +515,7 @@ OS_OBJ_QTY  OSSemPendAbort (OS_SEM  *p_sem,
         rt_ipc_pend_abort_1(&(p_sem->Sem.parent.suspend_thread));
     }
     
-    if(!opt&OS_OPT_POST_NO_SCHED)
+    if(!(opt&OS_OPT_POST_NO_SCHED))
     {
         rt_schedule();
     }