+++ rt_thread_3.1.3-ucosiii_3.03-wrapper/uCOS-III_Wrapper/uCOS-III/os_task.c
@@ -47,7 +47,6 @@
 ************************************************************************************************************************
 * Note(s)    : 1)由于RTT没有相关接口，因此以下函数没有实现
 *                   OSTaskQFlush
-*                   OSTaskQPendAbort
 *                   OSTaskChangePrio
 *                   OSTaskTimeQuantaSet
 ************************************************************************************************************************
@@ -623,6 +622,62 @@ CPU_BOOLEAN  OSTaskQPendAbort (OS_TCB  *p_tcb,
                                OS_OPT   opt,
                                OS_ERR  *p_err)
 {
+    OS_OPT _opt;
+    
+    CPU_SR_ALLOC();
+
+#ifdef OS_SAFETY_CRITICAL
+    if (p_err == (OS_ERR *)0) {
+        OS_SAFETY_CRITICAL_EXCEPTION();
+        return (DEF_FALSE);
+    }
+#endif
+
+#if OS_CFG_CALLED_FROM_ISR_CHK_EN > 0u
+    if (OSIntNestingCtr > (OS_NESTING_CTR)0) {              /* See if called from ISR ...                             */
+       *p_err = OS_ERR_PEND_ABORT_ISR;                      /* ... can't Pend Abort from an ISR                       */
+        return (DEF_FALSE);
+    }
+#endif
+
+#if OS_CFG_ARG_CHK_EN > 0u                                  /* ---------------- VALIDATE ARGUMENTS ------------------ */
+    switch (opt) {                                          /* User must supply a valid option                        */
+        case OS_OPT_POST_NONE:
+        case OS_OPT_POST_NO_SCHED:
+             break;
+
+        default:
+            *p_err = OS_ERR_OPT_INVALID;
+             return (DEF_FALSE);
+    }
+#endif
+    
+    CPU_CRITICAL_ENTER();
+#if OS_CFG_ARG_CHK_EN > 0u
+    if ((p_tcb == (OS_TCB *)0) ||                           /* Pend abort self?                                       */
+        (p_tcb == OSTCBCurPtr)) {
+        CPU_CRITICAL_EXIT();  
+       *p_err = OS_ERR_PEND_ABORT_SELF;                     /* ... doesn't make sense                                 */
+        return (DEF_FALSE);
+    }
+#endif
+//    if (p_tcb->PendOn != OS_TASK_PEND_ON_TASK_Q) {          /* Is task waiting for a message?                         */
+//        CPU_CRITICAL_EXIT();                                /* No                                                     */
+//       *p_err = OS_ERR_PEND_ABORT_NONE;
+//        return (DEF_FALSE);
+//    }    
+    CPU_CRITICAL_EXIT();  
+    
+    _opt = OS_OPT_PEND_ABORT_1 | opt;
+    OSQPendAbort(&p_tcb->MsgQ,_opt,p_err);
+    if(*p_err != OS_ERR_NONE)
+    {
+        return DEF_FALSE;
+    }   
+    else
+    {
+        return DEF_TRUE;
+    }    
 }
 #endif
 
@@ -1051,6 +1106,8 @@ CPU_BOOLEAN  OSTaskSemPendAbort (OS_TCB  *p_tcb,
 {
     OS_OPT _opt;
     
+    CPU_SR_ALLOC();
+    
 #ifdef OS_SAFETY_CRITICAL
     if (p_err == (OS_ERR *)0) {
         OS_SAFETY_CRITICAL_EXCEPTION();
@@ -1077,12 +1134,22 @@ CPU_BOOLEAN  OSTaskSemPendAbort (OS_TCB  *p_tcb,
     }
 #endif  
     
+    CPU_CRITICAL_ENTER();
+#if OS_CFG_ARG_CHK_EN > 0u    
     if ((p_tcb == (OS_TCB *)0) ||                           /* Pend abort self?                                       */
         (p_tcb == OSTCBCurPtr)) {
+        CPU_CRITICAL_EXIT(); 
        *p_err = OS_ERR_PEND_ABORT_SELF;
         return (DEF_FALSE);
     }
-      
+#endif
+//    if (p_tcb->PendOn != OS_TASK_PEND_ON_TASK_SEM) {        /* Is task waiting for a signal?                          */
+//        CPU_CRITICAL_EXIT();
+//       *p_err = OS_ERR_PEND_ABORT_NONE;
+//        return (DEF_FALSE);
+//    }
+    CPU_CRITICAL_EXIT(); 
+    
     _opt = OS_OPT_PEND_ABORT_1 | opt;
     OSSemPendAbort(&p_tcb->Sem,_opt,p_err);
     if(*p_err != OS_ERR_NONE)